Lldn frame (#1)
* testing something in git to familiarize with tool

* Create .gitignore

* Delete .tags_swap

* No changes in code were made

Added commentaries to better understandment of nwkDataReqTaskHandler and nwkTxTaskHandler

* Removed macSeq variable from Beacon Frame (NwkFrameBeaconHeader_t)

* Revert "Removed macSeq variable from Beacon Frame (NwkFrameBeaconHeader_t)"

This reverts commit 551e687a7d0b5d9d1ac364958ccc96059d9b1012.

* Revert "Revert "Removed macSeq variable from Beacon Frame (NwkFrameBeaconHeader_t)""

This reverts commit 546e957d40a19edb043ea62f413bf4c53f1c4a6b.

* Reversed changes in Beacon header

added more commenteries and undone changes

* fixed missing ;

* Update config.h

* add commentaries

* add commentaries

* Changed Beacon Header

* minor changes

* added coments and reverted some changes

* added comments

+++ libraries/LwMesh/src/nwk/nwkFrame.h
@@ -43,7 +43,7 @@
 /*
  * Copyright (c) 2014-2015 Atmel Corporation. All rights reserved.
  *
- * Licensed under Atmel's Limited License Agreement --> EULA.txt
+ * Licensed under Atmel's Limited Lic	ense Agreement --> EULA.txt
  */
 
 #ifndef _NWK_FRAME_H_
@@ -58,16 +58,16 @@ extern "C" {
 #include <sys\sysTypes.h>
 
 /*- Definitions ------------------------------------------------------------*/
-#define NWK_FRAME_MAX_PAYLOAD_SIZE   127
+#define NWK_FRAME_MAX_PAYLOAD_SIZE   127 // this impacts on phy.c file, see  phyRxBuffer[]
 
 /*- Types ------------------------------------------------------------------*/
 COMPILER_PACK_SET(1)
-typedef struct  NwkFrameHeader_t {
+typedef struct  NwkFrameHeader_t { // Size: 176 bits
 	uint16_t macFcf;
 	uint8_t macSeq;
 	uint16_t macDstPanId;
 	uint16_t macDstAddr;
-	uint16_t macSrcAddr;
+	uint16_t macSrcAddr; // Used for routing
 
 	struct {
 		uint8_t ackRequest : 1;
@@ -86,11 +86,18 @@ typedef struct  NwkFrameHeader_t {
 	};
 } NwkFrameHeader_t;
 
-typedef struct  NwkFrameBeaconHeader_t {
-	uint16_t macFcf;
-	uint8_t macSeq;
-	uint16_t macSrcPanId;
-	uint16_t macSrcAddr;
+typedef struct  NwkFrameBeaconHeader_t { // Size: 184 bits
+	uint8_t macFcf; // Frame Control Field:  changed size to LL standart
+	uint8_t macSeq; // Sequence Number: should only be avaible when Security is Enabled
+
+	// uint8_t macFlags;
+
+	/* Present in 802.15.4-2015 Beacon Frame Format */
+	uint8_t macSrcPanId;
+	uint8_t macSrcAddr;
+
+	// uint8_t macTimeSlot;
+
 	struct
 	{
 		uint16_t beaconOrder			: 4;
@@ -119,7 +126,7 @@ typedef struct NwkFrame_t {
 	union {
 		NwkFrameHeader_t header;
 		NwkFrameBeaconHeader_t beacon;
-		uint8_t data[NWK_FRAME_MAX_PAYLOAD_SIZE];
+		uint8_t data[NWK_FRAME_MAX_PAYLOAD_SIZE]; // Size: (8 * NWK_FRAME_MAX_PAYLOAD_SIZE) bits
 	};
 
 	uint8_t *payload;
